Remove TransferMsid check from Previos Oss number and add it to Previous Ioss number.
